tests: Add test cases for input advancing upon HasNext() call
The previous test cases assumed that the "input" advances on Next()
call. This is wrong. ChannelInput published in this repo is one example
of input advancing on HasNext() call rather than on Next() call.
The new test cases are covering this.

Note: the new cases are still failing; no fix yet.